Fold the two compression modules into one
The sequential compression trait and its parallel batch form lived in
separate files, but neither stands alone: the batch trait names the
sequential one as an associated type, and the adaptor exists only to
bridge them.

Merge both into crates/hash/src/compression.rs. The crate-root
re-exports are unchanged, so no caller outside the hash crate is
affected.

While merging, fix a stale doc reference to a trait that no longer
exists, replace two invented rustdoc sections with the batch
initialization contract that the Merkle tree builder actually relies on,
and drop an unused import. The fixed-input adaptor test becomes a
proptest against the sequential reference, alongside a boundary test.
